# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for GoPOS Backend",
        "title": "GoPOS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/shifts/active": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of all cashier shifts that are still open across the store (Admin Only).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Shifts"
                ],
                "summary": "Get all currently active/open shifts",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved active shifts",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/admin/shifts/{id}/force-close": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin can force close a shift that was left open by a cashier, calculating discrepancy and recording an audit log.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Shifts"
                ],
                "summary": "Force close a hanging cashier shift",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shift ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Actual Cash Data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ForceCloseInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shift successfully force-closed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticate user with email and password to receive a JWT access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User Authentication",
                "parameters": [
                    {
                        "description": "User Credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful with token payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid JSON payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "412": {
                        "description": "Account is inactive or disabled",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error while generating token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ForceCloseInput": {
            "type": "object",
            "required": [
                "total_cash_actual"
            ],
            "properties": {
                "total_cash_actual": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "models.LoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}